Refactor LLM service
