+++ app/jobs/create_llm_cover_letter_job.rb
@@ -2,6 +2,6 @@ class CreateLlmCoverLetterJob < ApplicationJob
   queue_as :llm
 
   def perform(job_application_id, temperature)
-    CoverLetterLlmService.new(job_application_id, temperature).create_cover_letter
+    Ai::CoverLetter.new(job_application_id, temperature).create_cover_letter
   end
 end
