+++ app/services/ai/llm_service.rb
@@ -1,0 +1,17 @@
+module Ai
+  class LlmService
+    def initialize(temperature)
+      @temperature = temperature.to_f
+      @client = OpenAI::Client.new(uri_base: Rails.application.config.openai[:url])
+    end
+
+    def llm_completions(prompt)
+      response = @client.completions(parameters: { model: Rails.application.config.openai[:model],
+                                                   prompt: prompt,
+                                                   max_tokens: Rails.application.config.openai[:max_tokens],
+                                                   temperature: @temperature })
+
+      response['choices']&.first&.dig('text')
+    end
+  end
+end
