+++ spec/jobs/create_llm_cover_letter_job_spec.rb
@@ -11,8 +11,8 @@ RSpec.describe CreateLlmCoverLetterJob do
     end
 
     it 'enqueues an event in the default queue' do
-      service = instance_double(CoverLetterLlmService, create_cover_letter: true)
-      allow(CoverLetterLlmService).to receive(:new).and_return(service)
+      service = instance_double(Ai::CoverLetter, create_cover_letter: true)
+      allow(Ai::CoverLetter).to receive(:new).and_return(service)
 
       described_class.perform_now(1, 0.5)
